+++ cedar-policy-core/src/ast.rs
@@ -22,6 +22,8 @@ mod entity;
 pub use entity::*;
 mod extension;
 pub use extension::*;
+mod id;
+pub use id::*;
 mod integer;
 pub use integer::{InputInteger, Integer};
 mod literal;
