+++ cedar-policy-core/src/ast/policy.rs
@@ -14,8 +14,7 @@
  * limitations under the License.
  */
 
-use crate::parser::err::ParseErrors;
-use crate::{ast::*, FromNormalizedStr};
+use crate::ast::*;
 use itertools::Itertools;
 use miette::Diagnostic;
 use serde::{Deserialize, Serialize};
@@ -787,96 +786,6 @@ impl From<StaticPolicy> for Arc<Template> {
     fn from(p: StaticPolicy) -> Self {
         let (t, _) = Template::link_static_policy(p);
         t
-    }
-}
-
-/// Like `Id`, except this specifically _can_ contain Cedar reserved identifiers.
-/// (It still can't contain, for instance, spaces or characters like '+'.)
-//
-// For now, internally, `AnyId`s are just owned `SmolString`s.
-#[derive(Serialize, Deserialize, Debug, PartialEq, Eq, Clone, Hash, PartialOrd, Ord)]
-pub struct AnyId(SmolStr);
-
-impl AnyId {
-    /// Create a new `AnyId` from a `String`, where it is the caller's
-    /// responsibility to ensure that the string is indeed a valid `AnyId`.
-    ///
-    /// When possible, callers should not use this, and instead use `s.parse()`,
-    /// which checks that `s` is a valid `AnyId`, and returns a parse error
-    /// if not.
-    ///
-    /// This method was created for the `From<cst::Ident> for AnyId` impl to use.
-    /// See notes on `Id::new_unchecked()`.
-    pub(crate) fn new_unchecked(s: impl Into<SmolStr>) -> AnyId {
-        AnyId(s.into())
-    }
-
-    /// Get the underlying string
-    pub fn into_smolstr(self) -> SmolStr {
-        self.0
-    }
-}
-
-impl AsRef<str> for AnyId {
-    fn as_ref(&self) -> &str {
-        &self.0
-    }
-}
-
-impl std::fmt::Display for AnyId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", &self.0)
-    }
-}
-
-// allow `.parse()` on a string to make an `AnyId`
-impl std::str::FromStr for AnyId {
-    type Err = ParseErrors;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        crate::parser::parse_anyid(s)
-    }
-}
-
-impl FromNormalizedStr for AnyId {
-    fn describe_self() -> &'static str {
-        "AnyId"
-    }
-}
-
-#[cfg(feature = "arbitrary")]
-impl<'a> arbitrary::Arbitrary<'a> for AnyId {
-    fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
-        // AnyId syntax:
-        // ['_''a'-'z''A'-'Z']['_''a'-'z''A'-'Z''0'-'9']*
-
-        let construct_list = |s: &str| s.chars().collect::<Vec<char>>();
-        let list_concat = |s1: &[char], s2: &[char]| [s1, s2].concat();
-        // the set of the first character of an AnyId
-        let head_letters = construct_list("ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz_");
-        // the set of the remaining characters of an AnyId
-        let tail_letters = list_concat(&construct_list("0123456789"), &head_letters);
-        // identifier character count minus 1
-        let remaining_length = u.int_in_range(0..=16)?;
-        let mut cs = vec![*u.choose(&head_letters)?];
-        cs.extend(
-            (0..remaining_length)
-                .map(|_| u.choose(&tail_letters))
-                .collect::<Result<Vec<&char>, _>>()?,
-        );
-        let s: String = cs.into_iter().collect();
-        crate::parser::parse_ident(&s).unwrap_or_else(|e| panic!("all strings constructed this way should be valid AnyIds, but this one is not: {s:?}\nerror: {e}"));
-        Ok(Self::new_unchecked(s))
-    }
-
-    fn size_hint(depth: usize) -> (usize, Option<usize>) {
-        arbitrary::size_hint::and_all(&[
-            // for arbitrary length
-            <usize as arbitrary::Arbitrary>::size_hint(depth),
-            // for arbitrary choices
-            // we use the size hint of a vector of `u8` to get an underestimate of bytes required by the sequence of choices.
-            <Vec<u8> as arbitrary::Arbitrary>::size_hint(depth),
-        ])
     }
 }
 
@@ -1748,7 +1657,7 @@ mod test {
 
     use super::{test_generators::*, *};
     use crate::{
-        ast::{entity, name, EntityUID},
+        ast::{entity, id, name, EntityUID},
         parser::{
             err::{ParseError, ParseErrors, ToASTError, ToASTErrorKind},
             parse_policy, Loc,
@@ -2039,7 +1948,7 @@ mod test {
     #[test]
     fn test_iter_once() {
         let id = EntityUID::from_components(
-            name::Name::unqualified_name(name::Id::new_unchecked("s")),
+            name::Name::unqualified_name(id::Id::new_unchecked("s")),
             entity::Eid::new("eid"),
         );
         let mut i = EntityIterator::One(&id);
@@ -2050,11 +1959,11 @@ mod test {
     #[test]
     fn test_iter_mult() {
         let id1 = EntityUID::from_components(
-            name::Name::unqualified_name(name::Id::new_unchecked("s")),
+            name::Name::unqualified_name(id::Id::new_unchecked("s")),
             entity::Eid::new("eid1"),
         );
         let id2 = EntityUID::from_components(
-            name::Name::unqualified_name(name::Id::new_unchecked("s")),
+            name::Name::unqualified_name(id::Id::new_unchecked("s")),
             entity::Eid::new("eid2"),
         );
         let v = vec![&id1, &id2];
